pendejiando en el frontend

+++ src/utils/utils.ts
@@ -2,7 +2,7 @@ export function formatDate(isoString: string): string {
   const date = new Date(isoString)
   const formatter = new Intl.DateTimeFormat('es-ES', {
     year: 'numeric',
-    month: 'long',
+    month: 'numeric',
     day: 'numeric'
   })
   return formatter.format(date)
